pyproject.toml
[tool.poetry]
name = "reta-lib-connector"
version = "0.1.0"
description = ""
authors = ["Oktapian <oktapian@jkt1.ebdesk.com>"]
readme = "Readme.md"
packages = [{include = "roubi"}]

[tool.poetry.dependencies]
python = "^3.10"
module-typica = {url = "https://github.com/oktapiancaw/module-typica/releases/download/v0.1.5-beta/module_typica-0.1.5.tar.gz"}
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
elasticsearch7 = "^7.17.9"
pika = "^1.3.2"
redis = "^5.0.1"
clickhouse-connect = "^0.7.0"
pymongo = "^4.6.1"
boto3 = "^1.34.34"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
